+++ client/src/main/scala/org/http4s/client/middleware/Logger.scala
@@ -5,7 +5,6 @@ package middleware
 import cats.effect._
 import fs2._
 import org.http4s.util.CaseInsensitiveString
-import org.log4s.{Logger ⇒ SLogger}
 
 /**
   * Simple Middleware for Logging All Requests and Responses
@@ -16,8 +15,8 @@ object Logger {
       logBody: Boolean,
       redactHeadersWhen: CaseInsensitiveString => Boolean = Headers.SensitiveHeaders.contains
   )(client: Client[F]): Client[F] =
-    ResponseLogger.apply0(logHeaders, logBody, redactHeadersWhen)(
-      RequestLogger.apply0(logHeaders, logBody, redactHeadersWhen)(
+    ResponseLogger.apply(logHeaders, logBody, redactHeadersWhen)(
+      RequestLogger.apply(logHeaders, logBody, redactHeadersWhen)(
         client
       )
     )
@@ -26,7 +25,7 @@ object Logger {
       logHeaders: Boolean,
       logBody: Boolean,
       redactHeadersWhen: CaseInsensitiveString => Boolean = Headers.SensitiveHeaders.contains)(
-      logger: SLogger)(implicit F: Effect[F]): F[Unit] = {
+      log: String => Unit)(implicit F: Effect[F]): F[Unit] = {
 
     val charset = message.charset
     val isBinary = message.contentType.exists(_.mediaType.binary)
@@ -68,7 +67,7 @@ object Logger {
     else {
       bodyText
         .map(body => s"$prelude $headers $body")
-        .map(text => logger.info(text))
+        .map(text => log(text))
         .compile
         .drain
     }