+++ client/src/main/scala/org/http4s/client/middleware/RequestLogger.scala
@@ -23,7 +23,8 @@ object RequestLogger {
   )(client: Client[F])(implicit ec: ExecutionContext = ExecutionContext.global): Client[F] =
     client.copy(open = Kleisli { req =>
       if (!logBody)
-        Logger.logMessage[F, Request[F]](req)(logHeaders, logBody)(logger) *> client.open(req)
+        Logger.logMessage[F, Request[F]](req)(logHeaders, logBody)(logger.info(_)) *> client.open(
+          req)
       else
         async.refOf[F, Vector[Segment[Byte, Unit]]](Vector.empty[Segment[Byte, Unit]]).flatMap {
           vec =>
@@ -40,7 +41,7 @@ object RequestLogger {
                   Logger.logMessage[F, Request[F]](req.withBodyStream(newBody))(
                     logHeaders,
                     logBody,
-                    redactHeadersWhen)(logger)
+                    redactHeadersWhen)(logger.info(_))
                 )
             )